separate location component into its own component, make fluid [#129]
